Split library-lifecycle ops out of manager config.rs
config.rs's module doc claimed config access, but the file also held
library-lifecycle operations that mutate the on-disk library presence and
the active-library pointer. Move rename_library, forget_encryption_key, and
forget_library verbatim into a new lifecycle.rs, leaving config.rs with the
genuine config-access surface (get_config, subscribe_config_changes,
set_pause_between_sides) plus the Discogs section.

Addresses #197 (lifecycle split). The lone per-field setter
(set_pause_between_sides) and the error-type inconsistency across the surface
are assessed in the review rather than churned here.
